+++ src/meeple/util/table_util.py
@@ -21,34 +21,31 @@ class ItemHeader(Enum):
 
 class CollectionHeader(Enum):
     NAME = ("Name", "name")
-    BOARDGAMES = ("Boardgames", "boardgames")
+    BOARDGAMES = ("Board Games", "boardgames")
     EXPANSIONS = ("Expansions", "expansions")
     UPDATED = ("Last Updated", "updated")
 
 
 def print_table(
-    rows: list,
-    headers: list = [],
-    lines: bool = False,
-    dim_border: bool = False,
-    zebra: bool = False,
+    rows: list, headers: list = [], dim_border: bool = False, zebra_rows: bool = False
 ) -> None:
-    row_styles = []
-    border_styles = []
-    if zebra:
-        row_styles = ["", "dim"]
+    border_styles, row_styles = [], []
     if dim_border:
         border_styles = "dim"
+    if zebra_rows:
+        row_styles = ["", "dim"]
+
     table = Table(
         box=box.ROUNDED,
         show_header=(len(headers) != 0),
-        show_lines=lines,
         row_styles=row_styles,
         border_style=border_styles,
     )
+
     for header in headers:
         table.add_column(header)
     for row in rows:
         table.add_row(*row)
+
     console = Console()
     console.print(table)